Extract binary detection in E2E tests
